+++ packages/react-magma-dom/src/components/Button/Button.stories.tsx
@@ -55,6 +55,11 @@ export default {
         options: ButtonType,
       },
     },
+    isLoading: {
+      control: {
+        type: 'boolean',
+      },
+    },
   },
 } as Meta;
 
@@ -71,6 +76,7 @@ Inverse.args = {
   ...Default.args,
   isInverse: true,
 };
+
 Inverse.decorators = [
   Story => (
     <Card background={magma.colors.foundation} isInverse>