+++ packages/react-magma-dom/src/components/DatePicker/index.tsx
@@ -182,9 +182,7 @@ export const DatePicker = React.forwardRef<HTMLInputElement, DatePickerProps>(
       if (inDateRange(newDate, convertedMinDate, convertedMaxDate)) {
         return newDate;
       } else if (convertedMaxDate || convertedMinDate) {
-        return isBefore(convertedMinDate, newDate)
-          ? convertedMinDate
-          : convertedMaxDate;
+        return convertedMinDate ? convertedMinDate : convertedMaxDate;
       }
     }