+++ packages/react-magma-dom/src/components/IconButton/IconButton.stories.tsx
@@ -61,6 +61,11 @@ export default {
         options: ButtonType,
       },
     },
+    isLoading: {
+      control: {
+        type: 'boolean',
+      },
+    },
   },
 } as Meta;