+++ packages/react-magma-dom/src/index.ts
@@ -72,6 +72,10 @@ export {
   DropdownMenuItemProps,
 } from './components/Dropdown/DropdownMenuItem';
 export {
+  DropdownMenuNavItem,
+  DropdownMenuNavItemProps,
+} from './components/Dropdown/DropdownMenuNavItem';
+export {
   DropdownSplitButton,
   DropdownSplitButtonProps,
 } from './components/Dropdown/DropdownSplitButton';
@@ -209,3 +213,4 @@ export { I18nContext } from './i18n';
 export { defaultI18n } from './i18n/default';
 export { I18nInterface } from './i18n/interface';
 export { useMediaQuery } from './hooks/useMediaQuery';
+export { InverseContext, useIsInverse } from './inverse';