+++ packages/react-magma-dom/CHANGELOG.md
@@ -3,6 +3,42 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+### 2.3.11-alpha.0 (2021-04-15)
+
+**Note:** Version bump only for package react-magma-dom
+
+### 2.3.10 (2021-04-14)
+
+**Note:** Version bump only for package react-magma-dom
+
+### 2.3.10-alpha.14 (2021-04-14)
+
+**Note:** Version bump only for package react-magma-dom
+
+### 2.3.10-alpha.13 (2021-04-13)
+
+**Note:** Version bump only for package react-magma-dom
+
+### 2.3.10-alpha.12 (2021-04-07)
+
+**Note:** Version bump only for package react-magma-dom
+
+### 2.3.10-alpha.11 (2021-04-07)
+
+**Note:** Version bump only for package react-magma-dom
+
+### 2.3.10-alpha.10 (2021-04-06)
+
+**Note:** Version bump only for package react-magma-dom
+
+### 2.3.10-alpha.9 (2021-04-06)
+
+**Note:** Version bump only for package react-magma-dom
+
+### 2.3.10-alpha.8 (2021-04-05)
+
+**Note:** Version bump only for package react-magma-dom
+
 ### 2.3.10-alpha.7 (2021-04-02)
 
 **Note:** Version bump only for package react-magma-dom